Fix > Image switch not working bug (#118)
The issue was caused by an external dropdown library that was causing
problems with the productDetails feature. I found that this library was
causing issues by listening for clicks on the page and changing the
select attributes, which triggered the mutation observer of the
productDetails. To fix the issue, I removed the external library and am
using the default select HTML until I can work on improving the styling.